requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.3
pyyaml==6.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
html2text==2020.1.16

# Data processing
pandas==2.1.4
numpy==1.24.3
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0

# Jupyter (optional)
ipykernel==6.27.1
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
